# kairos/openapi.yaml
openapi: 3.0.3
info:
  title: Kairos Energy Optimization API
  description: Two-endpoint API for Kairos home energy scheduling optimization. Use `/optimize` with physical inputs (device-specific parameters) or `/optimize-generic` with generic inputs (already converted parameters).
  version: 1.0.0
paths:
  /optimize:
    post:
      summary: Run Energy Schedule Optimization (Physical Inputs)
      description: Accepts device-specific physical parameters before conversion to the generic model. Supports multiple storage types (Home Battery, EV Battery, DHW Tank, Building Thermal Mass). The backend automatically converts these to the generic model and runs optimization.
      operationId: runOptimization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptimizationRequest'
      responses:
        '200':
          description: Optimization completed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizationResponse'
        '400':
          description: Invalid request payload or conversion error.
  /optimize-generic:
    post:
      summary: Run Energy Schedule Optimization (Generic Inputs)
      description: Accepts current states, constraints, and forecasts using generic inputs with generic asset classes (Sources, Loads, Storage). Recommended for integrated systems that have already converted device-specific physical parameters.
      operationId: runOptimizationGeneric
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericOptimizationRequest'
      responses:
        '200':
          description: Optimization completed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizationResponse'
        '400':
          description: Invalid request payload.
  /optimizations/latest:
    get:
      summary: Get Latest Optimization
      description: Returns the latest completed optimization and its original inputs. The record is held in memory and is cleared when the API process restarts.
      operationId: getLatestOptimization
      responses:
        '200':
          description: Latest optimization record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LatestOptimizationRecord'
        '404':
          description: No optimization has completed since the API started.
...
components:
  schemas:
...
    OptimizationRequest:
      type: object
      required:
        - timestamp
        - time_step_duration_hours
        - horizon_hours
        - grid
        - base_load
        - storage
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-09-30T14:00:05+02:00"
          description: ISO 8601 timestamp marking the start of the optimization period.
        time_step_duration_hours:
          type: number
          format: float
          example: 0.25
          description: Duration of each time step in hours (e.g., 0.25 for 15-minute intervals).
        horizon_hours:
          type: number
          format: float
          example: 24
          description: Planning horizon duration in hours (e.g., 24 for a day-ahead optimization).
        time_limit_s:
          type: integer
          default: 10
          minimum: 1
          maximum: 3600
          description: Maximum solver runtime in seconds.
        switching_penalty:
          type: number
          format: float
          default: 0.1
          minimum: 0.0
          description: Cost-equivalent penalty per storage mode change; set to 0 to disable.
        grid_peak_penalty:
          type: number
          format: float
          default: 0.1
          minimum: 0.0
          description: >-
            Dimensionless weight for the grid import/export peak tie-breaker. For example, at
            0.40 currency/kWh and 15-minute steps, a value of 0.1 means the optimizer can
            accept up to 1 cent higher energy cost in exchange for a 1 kW lower grid peak.
            Set to 0 to disable.
        grid:
          $ref: '#/components/schemas/Grid'
        pv:
          type: array
          items:
            $ref: '#/components/schemas/PV'
          description: PV systems (optional).
        base_load:
          $ref: '#/components/schemas/BaseLoad'
        controllable_loads:
          type: array
          items:
            $ref: '#/components/schemas/ControllableLoad'
          description: Schedulable appliances (optional).
        storage:
          type: array
          description: Storage assets specified with device-specific physical parameters.
          items:
            oneOf:
              - $ref: '#/components/schemas/HomeBattery'
              - $ref: '#/components/schemas/EVBattery'
              - $ref: '#/components/schemas/DHWTank'
              - $ref: '#/components/schemas/BuildingThermalMass'
            discriminator:
              propertyName: storage_type

    # --- SOURCES ---
    Grid:
      type: object
      required:
        - id
        - name
        - current_power
        - max_import_power
        - max_export_power
        - import_price_forecast
        - export_price_forecast
      properties:
        id:
          type: string
          example: "grid_1"
        name:
          type: string
          example: "Main Grid Connection"
        current_power:
          type: number
          description: Current power in Watts (+ for import, - for export).
          example: 1200.0
        max_import_power:
          type: number
          description: Max import power constraint [W].
          example: 11000.0
        max_export_power:
          type: number
          description: Max export power constraint [W].
          example: 11000.0
        import_price_forecast:
          type: array
          description: Import price forecast [price/Wh] per time step. Array has one entry per time step; each entry represents the average price for the period starting at that time step.
          items:
            type: number
        export_price_forecast:
          type: array
          description: Export price forecast [price/Wh] per time step. Array has one entry per time step; each entry represents the average price for the period starting at that time step.
          items:
            type: number
    PV:
      type: object
      required:
        - id
        - name
        - current_power
        - power_forecast
      properties:
        id:
          type: string
          example: "pv_1"
        name:
          type: string
          example: "Rooftop PV"
        current_power:
          type: number
          description: Current production power in Watts.
          example: 800.0
        power_forecast:
          type: array
          description: Power generation forecast [W] per time step. Array has one entry per time step; each entry represents the average power generated during the period starting at that time step.
          items:
            type: number

    # --- LOADS ---
    BaseLoad:
      type: object
      required:
        - id
        - name
        - current_power
        - power_forecast
      properties:
        id:
          type: string
          example: "base_load_household"
        name:
          type: string
          example: "Household Base Load"
        current_power:
          type: number
          description: Current consumption power in Watts (Always >= 0), measured as total house consumption minus all modelled devices.
          example: 450.0
        power_forecast:
          type: array
          description: Uncontrollable power demand forecast [W] per time step. Array has one entry per time step; each entry represents the average power consumed during the period starting at that time step.
          items:
            type: number
    ControllableLoad:
      type: object
      required:
        - id
        - name
        - current_power
        - average_power
        - energy_demand
      properties:
        id:
          type: string
          example: "dishwasher"
        name:
          type: string
          example: "Dishwasher"
        current_power:
          type: number
          description: Current consumption power in Watts (Always >= 0).
          example: 0.0
        average_power:
          type: number
          description: Average power [W].
          example: 1500.0
        energy_demand:
          type: number
          description: Total energy demand [Wh].
          example: 1800.0
        earliest_start_time:
          type: string
          format: date-time
          description: Earliest allowed start timestamp.
        latest_finish_time:
          type: string
          format: date-time
          description: Latest allowed finish timestamp.

    # --- PHYSICAL HOME BATTERY ---
    HomeBattery:
      type: object
      required:
        - id
        - name
        - storage_type
        - current_soc
        - energy_capacity
        - max_soc
        - min_soc
        - max_charge_power
        - max_discharge_power
        - charge_efficiency
        - discharge_efficiency
      properties:
        id:
          type: string
          example: "home_battery"
          description: Unique identifier for the storage asset.
        name:
          type: string
          example: "Home Battery"
          description: Name of the storage asset.
        storage_type:
          type: string
          enum: [home_battery]
          example: "home_battery"
          description: Physical storage type (home battery).
        current_soc:
          type: number
          minimum: 0.0
          maximum: 1.0
          example: 0.5
          description: Current state of charge as fraction [0-1].
        energy_capacity:
          type: number
          example: 10000.0
          description: Total battery capacity [Wh].
        max_soc:
          type: number
          minimum: 0.0
          maximum: 1.0
          example: 0.9
          description: Maximum state of charge for health protection [-].
        min_soc:
          type: number
          minimum: 0.0
          maximum: 1.0
          example: 0.1
          description: Minimum state of charge for health protection [-].
        max_charge_power:
          type: number
          example: 5000.0
          description: Maximum charging power [W].
        max_discharge_power:
          type: number
          example: 5000.0
          description: Maximum discharging power [W].
        charge_efficiency:
          type: number
          example: 0.95
          description: One-way charging efficiency [-].
        discharge_efficiency:
          type: number
          example: 0.95
          description: One-way discharging efficiency [-].
        passive_discharge_power:
          type: number
          example: 0.0
          description: Passive standby losses [W].

    # --- PHYSICAL EV BATTERY ---
    EVBattery:
      type: object
      required:
        - id
        - name
        - storage_type
        - current_soc
        - energy_capacity
        - min_soc
        - max_soc
        - max_charge_power
        - max_discharge_power
        - charge_efficiency
        - discharge_efficiency
        - vehicle_efficiency
        - round_trip_distance
        - expected_departure_time
        - expected_arrival_time
      properties:
        id:
          type: string
          example: "ev_battery"
          description: Unique identifier for the storage asset.
        name:
          type: string
          example: "EV Battery"
          description: Name of the storage asset.
        storage_type:
          type: string
          enum: [ev_battery]
          example: "ev_battery"
          description: Physical storage type (EV battery).
        current_soc:
          type: number
          minimum: 0.0
          maximum: 1.0
          example: 0.5
          description: Current state of charge as fraction [0-1].
        energy_capacity:
          type: number
          example: 60000.0
          description: Total battery capacity [Wh].
        max_soc:
          type: number
          minimum: 0.0
          maximum: 1.0
          example: 0.9
          description: Maximum state of charge for health protection [-].
        min_soc:
          type: number
          minimum: 0.0
          maximum: 1.0
          example: 0.1
          description: Minimum state of charge for health protection [-].
        max_charge_power:
          type: number
          example: 7400.0
          description: Maximum charging power [W].
        max_discharge_power:
          type: number
          example: 0.0
          description: Maximum discharging power [W] (Vehicle to Grid capability).
        charge_efficiency:
          type: number
          example: 0.95
          description: One-way charging efficiency [-].
        discharge_efficiency:
          type: number
          example: 0.95
          description: One-way discharging efficiency [-].
        passive_discharge_power:
          type: number
          example: 0.0
          description: Passive standby losses [W].
        vehicle_efficiency:
          type: number
          example: 0.005
          description: Vehicle efficiency [km/Wh].
        round_trip_distance:
          type: number
          example: 100.0
          description: Expected round trip distance [km].
        expected_departure_time:
          type: string
          format: date-time
          example: "2024-01-01T09:00:00Z"
          description: Expected vehicle departure time.
        expected_arrival_time:
          type: string
          format: date-time
          example: "2024-01-01T18:00:00Z"
          description: Expected vehicle arrival time (next opportunity to charge).

    # --- PHYSICAL DHW TANK ---
    DHWTank:
      type: object
      required:
        - id
        - name
        - storage_type
        - tank_volume
        - min_water_temperature
        - max_water_temperature
        - current_water_temperature
        - min_comfort_temperature
        - heat_loss_coefficient
        - heat_pump_electric_power
        - heat_pump_cop
        - morning_peak_energy_demand
        - evening_peak_energy_demand
        - morning_peak_time
        - evening_peak_time
      properties:
        id:
          type: string
          example: "dhw_tank"
          description: Unique identifier for the storage asset.
        name:
          type: string
          example: "DHW Tank"
          description: Name of the storage asset.
        storage_type:
          type: string
          enum: [dhw_tank]
          example: "dhw_tank"
        tank_volume:
          type: number
          example: 300.0
          description: Tank volume [liters].
        min_water_temperature:
          type: number
          example: 20.0
          description: Minimum safe water temperature [°C].
        max_water_temperature:
          type: number
          example: 60.0
          description: Maximum safe water temperature [°C].
        current_water_temperature:
          type: number
          example: 50.0
          description: Current water temperature [°C].
        min_comfort_temperature:
          type: number
          example: 40.0
          description: Minimum comfort temperature for supply [°C].
        heat_loss_coefficient:
          type: number
          example: 4.0
          description: Heat loss coefficient [W/°C].
        heat_pump_electric_power:
          type: number
          example: 3000.0
          description: Heat pump electric power [W].
        heat_pump_cop:
          type: number
          example: 3.0
          description: Heat pump coefficient of performance (COP) [-].
        morning_peak_energy_demand:
          type: number
          example: 150.0
          description: Typical morning peak DHW energy demand [Wh].
        evening_peak_energy_demand:
          type: number
          example: 180.0
          description: Typical evening peak DHW energy demand [Wh].
        morning_peak_time:
          type: string
          format: time
          example: "07:00"
          description: Time of morning peak demand [HH:MM].
        evening_peak_time:
          type: string
          format: time
          example: "19:00"
          description: Time of evening peak demand [HH:MM].

    # --- PHYSICAL BUILDING THERMAL MASS ---
    BuildingThermalMass:
      type: object
      required:
        - id
        - name
        - storage_type
        - floor_area
        - thermal_mass_coefficient
        - default_weather_compensation_temperature
        - current_indoor_temperature
        - max_comfort_temperature
        - heating_rate
        - cooldown_rate
        - heat_pump_cop_charge
        - heat_pump_cop_default
      properties:
        id:
          type: string
          example: "building_thermal_mass"
          description: Unique identifier for the storage asset.
        name:
          type: string
          example: "Building Thermal Mass"
          description: Name of the storage asset.
        storage_type:
          type: string
          enum: [building_thermal_mass]
          example: "building_thermal_mass"
        floor_area:
          type: number
          example: 100.0
          description: Building floor area [m²].
        thermal_mass_coefficient:
          type: number
          example: 200.0
          description: Thermal mass per unit area [Wh/m²/°C].
        default_weather_compensation_temperature:
          type: number
          example: 20.0
          description: Default weather compensation curve target temperature [°C].
        current_indoor_temperature:
          type: number
          example: 20.5
          description: Current indoor temperature [°C].
        max_comfort_temperature:
          type: number
          example: 21.0
          description: Maximum comfort temperature [°C].
        heating_rate:
          type: number
          example: 0.6
          description: Heating rate when heat pump set to +dT mode [°C/hour].
        cooldown_rate:
          type: number
          example: 0.1
          description: Cooldown rate when heat pump set to -dT mode [°C/hour].
        heat_pump_cop_charge:
          type: number
          example: 2.5
          description: Heat pump COP for charging (+dT mode) [-].
        heat_pump_cop_default:
          type: number
          example: 3.0
          description: Heat pump COP for default weather compensation mode [-].

    # --- RESPONSE ---
    OptimizationResponse:
      type: object
      required:
        - status
        - objective_cost
        - time_step_minutes
        - assets
      properties:
        status:
          type: string
          enum: [Optimal, Feasible, Infeasible]
          example: "Optimal"
        objective_cost:
          type: number
          description: Evaluated total objective cost over the planning horizon.
          example: 1.24
        time_step_minutes:
          type: number
          description: Duration of each scheduled interval in minutes.
          example: 15
        assets:
          type: object
          description: Schedules keyed by asset ID. Power schedule times mark interval starts; storage SoC schedule times mark interval ends.
          additionalProperties:
            $ref: '#/components/schemas/AssetSchedule'
    LatestOptimizationRecord:
      type: object
      required:
        - request
        - response
        - completed_at
        - solve_time_seconds
        - storage_metadata
      properties:
        request:
          type: object
          description: Original optimization request.
        response:
          $ref: '#/components/schemas/OptimizationResponse'
        completed_at:
          type: string
          format: date-time
        solve_time_seconds:
          type: number
        storage_metadata:
          type: object
          description: Storage capacity and SoC bounds needed by the dashboard.
          additionalProperties:
            type: object
            properties:
              energy_capacity:
                type: number
              min_soc:
                type: number
              max_soc:
                type: number
    AssetSchedule:
      type: object
      required:
        - setpoint
        - unit
        - schedule
      properties:
        setpoint:
          type: number
          description: Power setpoint for the first time step [W], equal to the first schedule value.
          example: 2500.0
        unit:
          type: string
          enum: [W]
          example: W
        schedule:
          type: array
          description: Scheduled power per time step, timestamped at each interval start.
          items:
            $ref: '#/components/schemas/SchedulePoint'
        soc_schedule:
          type: array
          description: Storage SoC normalized from 0.0 (empty) to 1.0 (full), with one point per interval timestamped at each interval end. The current SoC is not included.
          items:
            $ref: '#/components/schemas/SchedulePoint'
    SchedulePoint:
      type: object
      required:
        - time
        - value
      properties:
        time:
          type: string
          format: date-time
          description: ISO 8601 timestamp with timezone offset.
          example: "2026-10-05T14:00:00+02:00"
        value:
          type: number
          example: 1200.0

    # --- HEALTH RESPONSE ---
...
    GenericOptimizationRequest:
      type: object
      required:
        - timestamp
        - time_step_duration_hours
        - horizon_hours
        - grid
        - base_load
        - storage
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-09-30T14:00:05+02:00"
          description: ISO 8601 timestamp marking the start of the optimization period.
        time_step_duration_hours:
          type: number
          format: float
          example: 0.25
          description: Duration of each time step in hours (e.g., 0.25 for 15-minute intervals).
        horizon_hours:
          type: number
          format: float
          example: 24
          description: Planning horizon duration in hours (e.g., 24 for a day-ahead optimization).
        time_limit_s:
          type: integer
          default: 10
          minimum: 1
          maximum: 3600
          description: Maximum solver runtime in seconds.
        switching_penalty:
          type: number
          format: float
          default: 0.1
          minimum: 0.0
          description: Cost-equivalent penalty per storage mode change; set to 0 to disable.
        grid_peak_penalty:
          type: number
          format: float
          default: 0.1
          minimum: 0.0
          description: >-
            Dimensionless weight for the grid import/export peak tie-breaker. For example, at
            0.40 currency/kWh and 15-minute steps, a value of 0.1 means the optimizer can
            accept up to 1 cent higher energy cost in exchange for a 1 kW lower grid peak.
            Set to 0 to disable.
        grid:
          $ref: '#/components/schemas/Grid'
        pv:
          type: array
          items:
            $ref: '#/components/schemas/PV'
        base_load:
          $ref: '#/components/schemas/BaseLoad'
        controllable_loads:
          type: array
          items:
            $ref: '#/components/schemas/ControllableLoad'
        storage:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/GenericBattery'
              - $ref: '#/components/schemas/GenericDHWTank'
              - $ref: '#/components/schemas/GenericBuildingThermalMass'
          description: Storage assets specified with generic parameters.

    # --- GENERIC STORAGE ---

    # --- GENERIC BATTERY STORAGE (Home Battery, EV Battery) ---
    GenericBattery:
      type: object
      required:
        - id
        - name
        - storage_type
        - current_soc
        - energy_capacity
        - max_soc
        - min_soc
        - charge_efficiency
        - discharge_efficiency
        - passive_discharge_power
        - discharge_to_electrical_network
        - energy_demand_forecast
        - availability_window
        - max_charge_power
        - max_discharge_power
      properties:
        id:
          type: string
          example: "home_battery"
          description: Unique identifier for the storage asset.
        name:
          type: string
          example: "Home Battery"
          description: Name of the storage asset.
        storage_type:
          type: string
          enum: [battery]
          example: "battery"
          description: Storage class type (electrical battery storage).
        current_soc:
          type: number
          minimum: 0.0
          maximum: 1.0
          example: 0.5
          description: Current state of charge as fraction [0-1].
        energy_capacity:
          type: number
          description: Total capacity in Wh.
          example: 10000.0
        max_soc:
          type: number
          minimum: 0.0
          maximum: 1.0
          example: 0.9
          description: Maximum state of charge constraint for health protection [-].
        min_soc:
          type: number
          minimum: 0.0
          maximum: 1.0
          example: 0.1
          description: Minimum state of charge constraint for health protection [-].
        charge_efficiency:
          type: number
          example: 0.95
          description: One-way charging efficiency [-].
        discharge_efficiency:
          type: number
          example: 0.95
          description: One-way discharging efficiency [-].
        passive_discharge_power:
          type: number
          example: 0.0
          description: Passive standby losses [W].
        discharge_to_electrical_network:
          type: boolean
          example: true
          description: True for electrical storage, False for thermal storage.
        energy_demand_forecast:
          type: array
          description: Forecasted external energy demand [Wh] per time step (e.g., EV driving). Array has one entry per time step; each entry represents the energy demand during the period starting at that time step.
          items:
            type: number
        availability_window:
          type: array
          description: Availability vector per time step (1 = available, 0 = unavailable).
          items:
            type: integer
            enum: [0, 1]
        max_charge_power:
          type: number
          description: Maximum charging power [W].
          example: 5000.0
        max_discharge_power:
          type: number
          description: Maximum discharging power [W].
          example: 5000.0

    # --- GENERIC DHW TANK STORAGE ---
...